Move some unused code to a new unused.cpp file
There is lots of code in StrobeAlign that is unused. It is either commented
out or it exists as a function that is never called. Since I would like to
focus on the relevant code when working on the program, I would like to
somehow get rid of this code. I usually just delete old code, knowing that I
can find it again in the Git history if the need arises.

However, I can see that you may be reluctant to do that as it appears that
you sometimes run experiments by quickly commenting/uncommenting a function
call (for example, the choice of hash function). So to still have the code
somewhere more easily accessible, I suggest to move it to a file
"unused.cpp" for the moment. This PR adds that file and moves some unused
functions from index.cpp into it. Additionally, this adds an "unused-code"
target to CMake, which is only there to ensure that the code in "unused.cpp"
is compiled when running "make". This ensures that the functions are
runnable should they be needed.
